use float internally; fixed movement system
TODO: dynamic FPS? OpenGL?

+++ src/components/geometry.rs
@@ -5,18 +5,18 @@ use specs;
 ///
 /// 坐标单位为像素。
 pub struct Position {
-    pub x: i32,
-    pub y: i32,
-    pub z: i32,
+    pub x: f32,
+    pub y: f32,
+    pub z: f32,
 }
 
 impl Position {
-    pub fn new(x: i32, y: i32, z: i32) -> Position {
+    pub fn new(x: f32, y: f32, z: f32) -> Position {
         Position { x, y, z }
     }
 
-    pub fn new2(x: i32, y: i32) -> Position {
-        Position { x, y, z: 0 }
+    pub fn new2(x: f32, y: f32) -> Position {
+        Position { x, y, z: 0.0 }
     }
 }
